requirements.txt
aqtinstall
anki
PyQt5
pydocstyle
pylama